refactor: add config file to i18n
