+++ src/layouts/Layout.astro
@@ -1,6 +1,6 @@
 ---
 import { SITE } from "@config";
-import { DEFAULT_LOCALE } from "@i18n/locales";
+import { DEFAULT_LOCALE } from "@i18n/config";
 import { translateFor, getLocaleInfo } from "@i18n/utils";
 import "@styles/base.css";
 import getRelativePath from "@utils/getRelativePath";
